+++ v2/errors.go
@@ -35,10 +35,22 @@
 
 package envish
 
+import "fmt"
+
 // ErrEmptyKey is returned whenever we're given a key that is zero-length
 // or only contains whitespace
 type ErrEmptyKey struct{}
 
 func (e ErrEmptyKey) Error() string {
 	return "zero-length key, or key only contains whitespace"
+}
+
+// ErrNilPointer is returned whenever you call a method on the Env struct
+// with a nil pointer
+type ErrNilPointer struct {
+	method string
+}
+
+func (e ErrNilPointer) Error() string {
+	return fmt.Sprintf("nil pointer to environment store passed to %s", e.method)
 }